use cargo the normal way for the quickstart example, update docs and CI
