+++ .github/workflows/smoke-test.yml
@@ -117,24 +117,9 @@ jobs:
 
       - name: Test
         run: cargo test --verbose --all-features
-        
 
-  rust-examples:
-    runs-on: ubuntu-22.04
-    defaults:
-      run:
-        working-directory: docs/source
-    env:
-      CARGO_TERM_COLOR: always
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v3
-
-      - name: Install nightly
-        run: rustup update nightly
-
-      - name: Cargo run
-        run: cargo +nightly -Zscript getting-started/code/quickstart-framework.rs
+      - name: Run docs example
+        run: cd ../docs/source/getting-started/code/quickstart-framework-rust; cargo run
 
   python-test:
     needs: rust-build
