+++ docs/source/getting-started/code/quickstart-framework-rust/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "rust2"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+opendp = { path = "../../../../../rust/", features = ["contrib", "honest-but-curious"] }
+
+# Make sure the features are in sync with Cargo-for-docs.toml.
