+++ docs/source/getting-started/quickstart.rst
@@ -31,7 +31,7 @@ OpenDP is available for Python, R, and Rust.
 
         In a new directory run ``cargo init`` and then specify OpenDP as a dependency in ``Cargo.toml``:
 
-        .. literalinclude:: code/quickstart-framework.rs
+        .. literalinclude:: code/quickstart-framework-rust/Cargo-for-docs.toml
             :language: toml
             :start-after: init
             :end-before: /init
@@ -91,7 +91,7 @@ Let's apply Laplace noise to a value.
     .. tab-item:: Rust
         :sync: rust
 
-        .. literalinclude:: code/quickstart-framework.rs
+        .. literalinclude:: code/quickstart-framework-rust/src/main.rs
             :language: rust
             :start-after: demo
             :end-before: /demo
